requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
stripe==7.5.0
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.1.0
